edit: Allow tmux sessions to avoid using X11
Normally the DISPLAY variable is used to manage this but when
sharing a tmux session between machines the DISPLAY variable
sometimes points to a machine you are not sat in front off.

Using tmux session variables to control this allows it to be
overridden for the whole session rather than having to do
it seperately for each shell.

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>