tstamp: Make stopwatch work better when there is no newline
Currently if there is a partial line in the ouptut buffer then the '\r'
used to show the stopwatch causes us to overwrite the partial line. This
makes reading the logs difficult.

Fix this by using counted backspaces instead of '\r'. Whilst still not
perfect (e.g. when there is line wrap in the middle of the stopwatch
message) we will never damage the trace output.